Split api.py into a directory
